Cleanup and organize modules
- new top-level compile module
- validation.rs moves into compile
- break up compile.rs into smaller files
