+++ fea-rs/src/lib.rs
@@ -5,13 +5,11 @@ mod diagnostic;
 mod parse;
 mod token_tree;
 mod types;
-mod validate;
 
-pub use compile::compile;
+pub use compile::{compile, validate};
 pub use diagnostic::{Diagnostic, Level};
 pub use parse::grammar::root;
 pub use parse::util;
 pub use parse::{Kind, Parser, SyntaxError, TokenSet};
 pub use token_tree::{typed, AstSink, Node, NodeOrToken};
 pub use types::{GlyphMap, GlyphName};
-pub use validate::validate;
